+++ test/test_sanity_checks.py
@@ -15,10 +15,13 @@ class TestPowerLossSanityCheck(unittest.TestCase):
     def test_sanity_check(self, target_device_idx, xp):
         from specula.data_objects.simul_params import SimulParams
         from specula.processing_objects.power_loss import PowerLoss
+        from specula.processing_objects.atmo_propagation import AtmoPropagation
+        from specula.data_objects.source import Source
         sp = SimulParams(pixel_pupil=20, pixel_pitch=0.05)
-        obj = PowerLoss(simul_params=sp, wavelengthInNm=500.0, nd=2,
-                        prop_distance=400e3, receiver_diam=0.1,
-                        target_device_idx=target_device_idx)
+        on_axis_source = Source(polar_coordinates=[0,0], magnitude=0.0, height=600000, wavelengthInNm=500.0)
+        prop = AtmoPropagation(simul_params=sp, source_dict={'on_axis': on_axis_source}, doFresnel=True, upwards=True,
+                               wavelengthInNm=500.0, padding_factor=2)
+        obj = PowerLoss(simul_params=sp, prop=prop, target_device_idx=target_device_idx)
         obj.sanity_check()