+++ ChangeLog.md
@@ -41,6 +41,7 @@
 - Optimization of the compute\_ifs\_covmat function
 - Added Fraunhofer far field propagation
 - Fixed silent misparsing/confusing errors in split\_output() when an object, alias or output name contained a reserved '.', '-' or ':' character; added early validation of YAML section names in Simul
+- Updated calculation of power loss such that reference PSF also uses Fresnel propagation
 - ...
 
 ## [1.0.3] - 2026-05-18